+++ shop/cart/cart_modifiers_base.py
@@ -5,11 +5,53 @@ class BaseCartModifier(object):
     Price modifiers are the cart's counterpart to backends.
     It allows to implement Taxes and rebates / bulk prices in an elegant manner:
     
-    Everytime the cart is refreshed (via it's update() method), the cart will 
-    call all subclasses of this registered with their full path in the
-    settings.SHOP_CART_MODIFIERS setting.
+    Every time the cart is refreshed (via it's update() method), the cart will 
+    call all subclasses of this class registered with their full path in the
+    settings.SHOP_CART_MODIFIERS setting, calling methods defined here are 
+    in the following sequence:
+    
+    1. pre_process_cart: Totals are not computed, the cart is "rough": only 
+        relations and quantities are available
+    2. process_cart_item: Called for each cart_item in the cart. The current total
+        for this item is available as current_total
+    (2.a). get_extra_cart_item_price_field: A helper method provided for simple
+        use cases. It returns a tuple of (description, value) to add to the current
+        cart_item
+    3. process_cart: Called once for the whole cart. Here, all fields relative to 
+        cart items are filled, as well as the cart subtotal. The current total
+        is available as Cart.current_total (it includes modifications from previous
+        calls to this method, in other modifiers)
+    (3.a). get_extra_cart_price_field: A helper method for simple use cases. It 
+        returns a tuple of (description, value) to add to the current cart_item
+    4. post_process_cart: all totals are up-to-date, the cart is ready to be
+        displayed. Any change you make here must be consistent!
     """
+    
+    #===========================================================================
+    # Processing hooks
+    #===========================================================================
+    
+    def pre_process_cart(self, cart, state):
+        """
+        This method will be called before the cart starts being processed.
+        Totals are not updated yep (obviously), but this method can be useful to 
+        gather some information on products in the cart.
         
+        The `state` parameter is further passed to process_cart_item, process_cart,
+        and post_process_cart, so it can be used as a way to store per-request
+        arbitrary information.
+        """
+        pass
+    
+    def post_process_cart(self, cart, state):
+        """
+        This method will be called after the cart was processed.
+        The Cart object is "final" and all the fields are computed. Remember that
+        anything changed at this point should be consistent: if updating the price
+        you should also update all relevant totals (for example).
+        """
+        pass
+    
     def process_cart_item(self, cart_item, state):
         """
         This will be called for every line item in the Cart:
@@ -53,7 +95,11 @@ class BaseCartModifier(object):
             cart.current_total = cart.current_total + price
             cart.extra_price_fields.append(field)
         return cart
-    
+
+    #===========================================================================
+    # Simple methods
+    #===========================================================================
+
     def get_extra_cart_item_price_field(self, cart_item):
         """
         Get an extra price field tuple for the current cart_item: